% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processControls.R
\name{defineControls}
\alias{defineControls}
\title{Identify the controls in a MIDAS file}
\usage{
defineControls(midas_file, control_condition, save_file = FALSE)
}
\arguments{
\item{midas_file}{A MIDAS matrix or filename, or a vector of filenames}

\item{control_condition}{A string depicting the control condition. Each control treatment has to be separated by a '+'.}

\item{save_file}{Whether the resulting MIDAS matrix should be saved in a file}
}
\description{
Rename the ID:type column of a MIDAS file which correspond to control treatment
}
\examples{
defineControls("blunt_MIDAS.csv", "BSA+DMSO")
}
