% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMIDAS.R
\name{filterMIDAS}
\alias{filterMIDAS}
\title{Filter out entries from a MIDAS}
\usage{
filterMIDAS(midas_file, filters, save_file = FALSE)
}
\arguments{
\item{midas_file}{A MIDAS matrix or filename}

\item{filters}{A list of filters. A filter is of the form READOUT, PERT which means that the entries for PERT or READOUT are removed, or of the form READOUT@PERT, in which case the entries for READOUT when perturbation PERT is present are removed}

\item{save_file}{Whether the resulting MIDAS matrix should be saved in a file}
}
\description{
Filter out some entries from a MIDAS file by setting them to NA
}
\examples{
filterMIDAS blunt_MIDAS.csv AKT@AKTi
}
\seealso{
extractMIDAS, defineControls
NOTE: this script assumes that perturbation and readout names are distinct
}
