% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processControls.R
\name{normaliseByControls}
\alias{normaliseByControls}
\title{Normalise by a control condition}
\usage{
normaliseByControls(midas_file, control_condition, save_file = FALSE)
}
\arguments{
\item{midas_file}{A MIDAS matrix or filename, or a vector of filenames}

\item{control_condition}{A string depicting the control condition. Each control treatment has to be separated by a '+'.}

\item{save_file}{Whether the resulting MIDAS matrix should be saved in a file}
}
\description{
Normalise all values by the respective control and renames the ID:type column of a midas file corresponding to control treatment
}
\examples{
normaliseByControls("blunt_MIDAS.csv", "BSA+DMSO") # BSA and DMSO are the controls used for the inhibitors and ligands respectively
}
