\name{processMIDAS-package}
\alias{processMIDAS-package}
\alias{processMIDAS}
\docType{package}
\title{
\packageTitle{processMIDAS}
}
\description{
\packageDescription{processMIDAS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{processMIDAS}
\packageIndices{processMIDAS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{processMIDAS}

Maintainer: \packageMaintainer{processMIDAS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
